backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==2.1.0
anthropic==0.7.7
openai==1.3.5
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic==2.5.0
asyncio==3.4.3
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1